+++ src/lib/export/export.ts
@@ -48,19 +48,3 @@ export async function executeExport(
     }
   }
 }
-
-// ─── Shared helpers ───────────────────────────────────────────────
-
-export function downloadFile(content: string, filename: string, mimeType: string): void {
-  const blob = new Blob([content], { type: mimeType });
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement("a");
-  a.href = url;
-  a.download = filename;
-  a.click();
-  URL.revokeObjectURL(url);
-}
-
-export function today(): string {
-  return new Date().toISOString().slice(0, 10);
-}
