+++ tests/export.test.ts
@@ -2,8 +2,8 @@ import { describe, it, expect, beforeAll, afterAll } from "vitest";
 import { buildTallies } from "../src/lib/agg/buildTallies";
 import type { Question, Tally } from "../src/lib/agg/types";
 import { setupDuckDB, teardownDuckDB } from "./helpers/duckdb";
-import { buildExportGrids } from "../src/lib/export/exportGrid";
-import { talliesToLongRows } from "../src/lib/export/longFormat";
+import { buildExportGrids } from "../src/lib/export/formatters/grid";
+import { talliesToLongRows } from "../src/lib/export/formatters/longFormat";
 import { formatCSV } from "../src/lib/export/formatters/csv";
 import { formatTSV } from "../src/lib/export/formatters/tsv";
 import { formatMarkdown } from "../src/lib/export/formatters/markdown";
